Add files via upload
I also included the Xbox 360 version of the unpacker for Halo Spartan Strike(360 version). I wasn't able to test if it works for the Xbox 360 version of the Halo Spartan Assault but it does for Halo Spartan Strike's 460 version. To use the unpackers, place them in the same directory as the folder that stores the BlueResource container files. You will need Python 3 installed.